refactor: MemoInput への変換処理を分けた
