+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/model/MemoInputConvertor.kt
@@ -1,0 +1,10 @@
+package app.kaito_dogi.mybrary.core.supabase.model
+
+import app.kaito_dogi.mybrary.core.data.command.PostMemoCommand
+
+internal fun PostMemoCommand.toInput() = MemoInput(
+  myBookId = this.myBookId,
+  content = this.content,
+  startPage = this.startPage,
+  endPage = this.endPage,
+)
